Introduce URXVT_TAB environment variable
Process in each tab will now have a URXVT_TAB variable set in their
environment.  It’s value is a total number of tabs opened by urxvt
followed by a colon.  The number starts at zero and always increases.

The colon is intended for future extension as more information may be
passed over time in the variable.

This feature can be used to provide per-tab initialisation or command.
For example, ~/.bashrc could include:

    if [ -n "$URXVT_TAB" ]; then
        tab=$URXVT_TAB
        unset URXVT_TAB

        set_urxvt_tab_name() {
            printf '\033]777;tabbedx;set_tab_name;%s\007' "$1"
        }

        case "$tab" in
        0:*)
            set_urxvt_tab_name emacs
            exec emacs
            ;;
        1:*)
            set_urxvt_tab_name work
            ;;
        2:*)
            set_urxvt_tab_name nethack
            exec nethack
            ;;
        esac
    fi

to name the first three tabs ‘emacs’, ‘work’ and ‘nethack’ and start
Emacs and NetHack in their respective tabs.